requirements.txt
Flask>=2.3,<3.0
openpyxl>=3.1,<4.0
psycopg2-binary
python-dotenv
